fix: Вынес фильтрацию по сумме в отдельный компонент
