+++ src/pages/GlobalCategories/globalCategories.module.scss
@@ -32,68 +32,6 @@ main {
       padding: 0;
       color: $text-silver;
     }
-    .priceSection {
-      display: flex;
-      flex-direction: column;
-      .price {
-        font-size: 22px;
-        padding: 20px 32px;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        border-top: 1px solid $lines;
-        border-bottom: 1px solid $lines;
-        cursor: pointer;
-        .norotated {
-          transition: 0.4s;
-        }
-        .rotated {
-          transition: 0.4s;
-          transform: rotate(0.25turn);
-        }
-      }
-      .noslidecontainer {
-        height: 0;
-        transition: 0.5s;
-        opacity: 0;
-        .inputs {
-          opacity: 0;
-          transform: 0.5s;
-        }
-      }
-      .slidecontainer {
-        height: auto;
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        flex-wrap: nowrap;
-        opacity: 1;
-        transition: 0.5s;
-        .inputs {
-          opacity: 1;
-          min-height: 60px;
-          display: flex;
-          justify-content: center;
-          flex-direction: column;
-          padding: 10px 20px;
-          div {
-            display: flex;
-            justify-content: flex-start;
-            &:first-child {
-              margin-bottom: 15px;
-            }
-            span {
-              margin-right: 10px;
-            }
-            input {
-              font-style: italic;
-              border: none;
-              border-bottom: 1px solid $lines;
-            }
-          }
-        }
-      }
-    }
   }
   .products {
     width: 95%;
